requirements.txt
Django==4.1.6
djangorestframework==3.14.0
psycopg2>=2.8.6,<2.9
Pillow>=8.2.0,<8.3.0
flake8==6.0.0
drf-yasg==1.21.5